Initial command restructure
Prior to adding 'create/get/delete nodegroup' some refactoring
has been done on the commands. We had a single file for all
the commands for that verb so this has been changed to separate
folders/files.
