+++ pkg/commands/common.go
@@ -1,0 +1,20 @@
+package commands
+
+import (
+	"os"
+	"strings"
+
+	"github.com/kubicorn/kubicorn/pkg/logger"
+)
+
+// GetNameArg tests to ensure there is only 1 name argument
+func GetNameArg(args []string) string {
+	if len(args) > 1 {
+		logger.Critical("only one argument is allowed to be used as a name")
+		os.Exit(1)
+	}
+	if len(args) == 1 {
+		return (strings.TrimSpace(args[0]))
+	}
+	return ""
+}
